@@ -7,14 +7,14 @@ class Admin < ApplicationRecord
   devise :omniauthable, :omniauth_providers => [:google]
 
   def self.from_omniauth(access_token)
-    where(provider: 'google', provider_id: access_token.extra.raw_info.id).first_or_create do |admin|
+    where( provider: 'google',
+           provider_id: access_token.extra.raw_info.id )
+    .first_or_create do |admin|
       admin.email = access_token.info.email
     end
   end
 
   has_many :rubrics, foreign_key: :created_by
-  has_many :rubrics, foreign_key: :closed_by
-  has_many :rubrics, foreign_key: :reviwee
 
   def admin?
     true
