@@ -1,5 +1,5 @@
 class HomePolicy < ApplicationPolicy
   def index?
-    true # Open to public
+    true
   end
 end
