@@ -1,7 +1,7 @@
 RSpec.describe HomeController, type: :routing do
-  describe "routing" do
-    xit "routes to #index" do
-      expect(get: "/").to route_to("home#index")
+  describe 'routing' do
+    it 'routes to #index' do
+      expect(get: '/').to route_to('home#index')
     end
   end
 end
